skillgap/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
click==8.1.7
PyPDF2==3.0.1
python-docx==1.1.0


